Fix ladybug securityfilterchain
