+++ core/src/main/resources/AppConstants.properties
@@ -569,8 +569,9 @@ ibistesttool.directory=${log.dir}/testtool
 ## (default is ${jdbc.migrator.active} when available, otherwise default is false)
 #ladybug.jdbc.migrator.active=true
 
-## DataSource to be used by Ladybug, and it's Liquibase script (default is ${jdbc.datasource.default})
-#ladybug.jdbc.datasource=
+## DataSource to be used by Ladybug, and it's Liquibase script.
+## When set the ladybug will use the database to store it's reports.
+ladybug.jdbc.datasource=
 
 ## Enable report generator in all environments (default: false when dtap.stage=ACC or dtap.stage=PRD, otherwise true)
 #testtool.enabled=true
